requirements.txt
flask==3.0.0
flask-cors==4.0.0
boto3==1.34.34
msal==1.26.0
pyjwt==2.8.0
cryptography==42.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
